Add even more general form component
